aplicando estructura prolija al proyecto y borrando archivos que no se usan
