+++ common/components/ui/alert-dialog.tsx
@@ -3,8 +3,8 @@
 import * as React from "react"
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog"
 
-import { cn } from "@/lib/utils"
-import { buttonVariants } from "@/components/ui/button"
+import { cn } from "@/common/utils/utils"
+import { buttonVariants } from "@/common/components/ui/button"
 
 function AlertDialog({
   ...props
