+++ common/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/common/utils/utils"
 
 function Input({ className, type, ...props }: React.ComponentProps<"input">) {
   return (
