+++ common/hooks/use-mobile.ts
